+++ packages/client/types/constEnums.ts
@@ -55,7 +55,7 @@ export enum Card {
 }
 
 export enum ContactInfo {
-  TELEPHONE = '612-227-5673',
+  TELEPHONE = '+1 772-519-7250',
   EMAIL_BILLING = 'billing@parabol.co',
   EMAIL_LOVE = 'love@parabol.co'
 }
@@ -347,11 +347,6 @@ export enum RightSidebar {
 export enum Security {
   SALT_ROUNDS = 12,
   MASS_INVITATION_TOKEN_LENGTH = 12
-}
-
-export enum ServerChannel {
-  GQL_EXECUTOR_STREAM = 'gqlStream',
-  GQL_EXECUTOR_CONSUMER_GROUP = 'gqlConsumerGroup'
 }
 
 export enum SubscriptionChannel {