+++ packages/server/HocusPocusWebSocket.ts
@@ -1,11 +1,13 @@
 import type {WebSocket} from 'uWebSockets.js'
 import {EventEmitter} from 'tseep'
+import type {HocusPocusSocketData} from './hocusPocusHandler'
+import {Logger} from './utils/Logger'
 
 // This is a mock WebSocket that fills as much of the standard WebSocket contract that HocusPocus requires
 export class HocusPocusWebSocket extends EventEmitter {
-  private ws: WebSocket<any>
+  private ws: WebSocket<HocusPocusSocketData>
   readyState = 1
-  constructor(ws: WebSocket<any>) {
+  constructor(ws: WebSocket<HocusPocusSocketData>) {
     super()
     this.ws = ws
     this.on('close', () => {
@@ -13,15 +15,31 @@ export class HocusPocusWebSocket extends EventEmitter {
     })
   }
   close(code?: number, reason?: string) {
-    if (this.readyState !== 1) return
-    this.ws.end(code, reason)
+    if (this.ws.getUserData().closed) return
+    this.readyState = 3
+    try {
+      this.ws.end(code, reason)
+    } catch (e) {
+      Logger.error({closed: this.ws.getUserData().closed, readyState: this.readyState})
+      Logger.error(e)
+    }
   }
   ping() {
-    if (this.readyState !== 1) return
-    this.ws.ping()
+    if (this.ws.getUserData().closed) return
+    try {
+      this.ws.ping()
+    } catch (e) {
+      Logger.error({closed: this.ws.getUserData().closed, readyState: this.readyState})
+      Logger.error(e)
+    }
   }
   send(message: Uint8Array) {
-    if (this.readyState !== 1) return
-    this.ws.send(message, true)
+    if (this.ws.getUserData().closed) return
+    try {
+      this.ws.send(message, true)
+    } catch (e) {
+      Logger.error({closed: this.ws.getUserData().closed, readyState: this.readyState})
+      Logger.error(e)
+    }
   }
 }