+++ packages/server/hocusPocus.ts
@@ -6,8 +6,9 @@ import {TiptapTransformer} from '@hocuspocus/transformer'
 import type {JSONContent} from '@tiptap/core'
 import StarterKit from '@tiptap/starter-kit'
 import tracer from 'dd-trace'
+import ms from 'ms'
 import {pack, unpack} from 'msgpackr'
-import {encodeStateAsUpdate} from 'yjs'
+import {Doc, encodeStateAsUpdate} from 'yjs'
 import {getNewDataLoader} from './dataloader/getNewDataLoader'
 import getKysely from './postgres/getKysely'
 import type {Pageroleenum} from './postgres/types/pg'
@@ -35,8 +36,24 @@ const pushGQLTitleUpdates = async (pageId: number) => {
   dataLoader.dispose()
 }
 
-type Req = IncomingMessage & {userId?: string}
+type Req = IncomingMessage & {userId?: string; tms: string[]}
 
+const getMeetingTeamId = async (meetingId: string) => {
+  // This is a top 15 query in terms of total time because of call frequency, so caching is necessary
+  const redis = getRedis()
+  const key = `meetingTeamId:${meetingId}`
+  const cachedTeamId = await redis.get(key)
+  if (cachedTeamId) return cachedTeamId
+  const meeting = await getKysely()
+    .selectFrom('NewMeeting')
+    .select('teamId')
+    .where('id', '=', meetingId)
+    .executeTakeFirst()
+  if (!meeting) return null
+  const {teamId} = meeting
+  redis.set(key, teamId, 'PX', ms('3h')).catch(() => {})
+  return teamId
+}
 export const redisHocusPocus = new RedisServerAffinity({
   pack,
   redis: getRedis(),
@@ -53,7 +70,6 @@ export const hocuspocus = new Hocuspocus({
     const authToken = getVerifiedAuthToken(authTokenStr)
     const userId = authToken?.sub
     if (authTokenStr && !userId) {
-      // If there _is_ an authToken & it is bad (expired, invalid) log out
       const err = new Error('Unauthenticated')
       ;(err as any).reason = 'Unauthenticated'
       throw err
@@ -61,12 +77,44 @@ export const hocuspocus = new Hocuspocus({
     // Unauthenticated users are allowed for public pages
     // put the userId on the request because context isn't available until onAuthenticate
     request.userId = authToken?.sub
+    request.tms = authToken?.tms ?? []
   },
 
   async onAuthenticate(data) {
     const {documentName, request, connectionConfig} = data
     const userId = (request as Req).userId
+    if (documentName.startsWith('meeting:')) {
+      const tms = (request as Req).tms
+      const [, meetingId] = documentName.split(':')
+      if (!meetingId) throw new Error(`Invalid meetingId: ${meetingId}`)
+      if (!userId) throw new Error(`Meeting awareness requires a userId`)
+      const meetingTeamId = await getMeetingTeamId(meetingId)
+      if (!tms.includes(meetingTeamId ?? '')) {
+        // They may have a stale token or client is making a request it shouldn't.
+        // Unsure how to reproduce, but I see errors in DD logs.
+        const teamMember = await getKysely()
+          .selectFrom('TeamMember')
+          .select('id')
+          .where('teamId', '=', meetingTeamId)
+          .where('userId', '=', userId)
+          .where('isNotRemoved', '=', true)
+          .executeTakeFirst()
+        if (!teamMember) {
+          const error = new Error(
+            `Meeting awareness requires being on the team: ${userId} ${meetingTeamId}`
+          )
+          ;(error as any).reason = 'Unauthorized'
+          throw error
+        }
+      }
+      return {userId}
+    }
     const [dbId] = CipherId.fromClient(documentName)
+    if (dbId === 0) {
+      const error = new Error(`Invalid document request from client: ${documentName}`)
+      logError(error, {userId, tags: {dbId, documentName}})
+      throw error
+    }
     let pageAccess: {role: Pageroleenum} | undefined
     if (userId) {
       pageAccess = await getKysely()
@@ -85,7 +133,11 @@ export const hocuspocus = new Hocuspocus({
         .limit(1)
         .executeTakeFirst()
       if (!pageAccess) {
-        throw new Error(`Document does not exist or user is not authorized: ${dbId}`)
+        const error = new Error(
+          `Document does not exist or user is not authorized: ${documentName}`
+        )
+        logError(error, {userId, extras: {dbId}})
+        throw error
       }
     }
     const {role} = pageAccess
@@ -99,7 +151,10 @@ export const hocuspocus = new Hocuspocus({
     new Database({
       // Return a Promise to retrieve data …
       fetch: async ({documentName}) => {
-        const [dbId] = CipherId.fromClient(documentName)
+        const [dbId, , entity] = CipherId.fromClient(documentName)
+        if (entity === 'meeting') {
+          return Buffer.from(encodeStateAsUpdate(new Doc()))
+        }
         const pg = getKysely()
         const res = await pg
           .selectFrom('Page')
@@ -120,7 +175,8 @@ export const hocuspocus = new Hocuspocus({
         return Buffer.from(encodeStateAsUpdate(yDoc))
       },
       store: async ({documentName, state, document}) => {
-        const [dbId] = CipherId.fromClient(documentName)
+        const [dbId, , entity] = CipherId.fromClient(documentName)
+        if (entity === 'meeting') return
         // TODO: don't transform the document into content. just traverse the yjs doc for speed
         const content = TiptapTransformer.fromYdoc(document, 'default') as JSONContent
         const {updatedTitle} = await tracer.trace(
@@ -173,7 +229,6 @@ hocuspocus.storeDocumentHooks = (
           })
         })
         .catch((error) => {
-          // normally hocuspocus would throw here, but we just want to log the error and continue on
           logError(error, {tags: {documentName: document.name}})
           // TODO report error to client so it can retry saving